refactor(orbiter): move state related modules
