+++ src/orbiter/src/serializers/bounded.rs
@@ -11,7 +11,7 @@ use crate::serializers::utils::{
     bytes_to_short_string, bytes_to_string, key_to_bytes, long_string_to_bytes, metadata_to_bytes,
     principal_to_bytes, short_string_to_bytes, string_to_bytes,
 };
-use crate::types::state::{
+use crate::state::types::state::{
     AnalyticKey, AnalyticSatelliteKey, PageView, PageViewDevice, TrackEvent,
 };
 use std::borrow::Cow;
